refactor(recurrence): split service.go below 1000 lines
Move event, todo, and journal list helpers into sibling files.
No behavior change.

Related: #652
