+++ internal/recurrence/service.go
@@ -3,18 +3,13 @@ package recurrence
 import (
 	"context"
 	"database/sql"
-	"sort"
 	"strings"
 	"time"
 
 	"github.com/teambition/rrule-go"
 
-	"github.com/douglasdemoura/chroncal/internal/event"
-	"github.com/douglasdemoura/chroncal/internal/journal"
-	"github.com/douglasdemoura/chroncal/internal/model"
 	"github.com/douglasdemoura/chroncal/internal/storage"
 	"github.com/douglasdemoura/chroncal/internal/timeutil"
-	"github.com/douglasdemoura/chroncal/internal/todo"
 )
 
 // Service handles recurrence expansion and the cache.
@@ -235,69 +230,6 @@ func newRRuleSet(rule, tz string, dtstart time.Time, dur time.Duration, exDates,
 	return rruleSet{set: set, loc: loc, rdateSet: buildRDateSet(rDates), dur: dur}, true
 }
 
-// eventOverridesByUID fetches every override for the given recurring masters in
-// a single query and groups them by master UID. There is no SELECT per master.
-// A failed fetch is propagated, not swallowed. A render of a master as if it
-// had no overrides would show a stale, un-overridden series in silence.
-func (s *Service) eventOverridesByUID(ctx context.Context, masters []storage.Event) (map[string][]storage.Event, error) {
-	if len(masters) == 0 {
-		return nil, nil
-	}
-	uids := make([]string, len(masters))
-	for i, m := range masters {
-		uids[i] = m.Uid
-	}
-	rows, err := s.q.ListOverridesByUIDs(ctx, uids)
-	if err != nil {
-		return nil, err
-	}
-	byUID := make(map[string][]storage.Event, len(rows))
-	for _, r := range rows {
-		byUID[r.Uid] = append(byUID[r.Uid], r)
-	}
-	return byUID, nil
-}
-
-// todoOverridesByUID is the todo analogue of eventOverridesByUID.
-func (s *Service) todoOverridesByUID(ctx context.Context, masters []storage.Todo) (map[string][]storage.Todo, error) {
-	if len(masters) == 0 {
-		return nil, nil
-	}
-	uids := make([]string, len(masters))
-	for i, m := range masters {
-		uids[i] = m.Uid
-	}
-	rows, err := s.q.ListTodoOverridesByUIDs(ctx, uids)
-	if err != nil {
-		return nil, err
-	}
-	byUID := make(map[string][]storage.Todo, len(rows))
-	for _, r := range rows {
-		byUID[r.Uid] = append(byUID[r.Uid], r)
-	}
-	return byUID, nil
-}
-
-// journalOverridesByUID is the journal analogue of eventOverridesByUID.
-func (s *Service) journalOverridesByUID(ctx context.Context, masters []storage.Journal) (map[string][]storage.Journal, error) {
-	if len(masters) == 0 {
-		return nil, nil
-	}
-	uids := make([]string, len(masters))
-	for i, m := range masters {
-		uids[i] = m.Uid
-	}
-	rows, err := s.q.ListJournalOverridesByUIDs(ctx, uids)
-	if err != nil {
-		return nil, err
-	}
-	byUID := make(map[string][]storage.Journal, len(rows))
-	for _, r := range rows {
-		byUID[r.Uid] = append(byUID[r.Uid], r)
-	}
-	return byUID, nil
-}
-
 // cancelledRecurringMaster reports whether the row is a recurring master that
 // is cancelled. A cancel of a recurring series cancels the whole series.
 // Such a master expands to no occurrences.
@@ -363,201 +295,6 @@ func (rs rruleSet) between(from, to time.Time) []time.Time {
 		}
 	}
 	return kept
-}
-
-// newEventRRuleSet parses evt's RRULE into a reusable rruleSet. See newRRuleSet
-// for includeExDates.
-func newEventRRuleSet(evt event.Event, includeExDates bool) (rruleSet, bool) {
-	return newRRuleSet(evt.RecurrenceRule, evt.Timezone, evt.StartTime,
-		evt.EndTime.Sub(evt.StartTime), evt.ParseExDates(), evt.ParseRDates(), includeExDates)
-}
-
-// newEventOccChecker builds a reusable orphan-detection checker for a recurring
-// event master (EXDATEs ignored, so an EXDATE never hides a legitimate override).
-// A cancelled series has no occurrences, so its checker matches nothing — this is
-// what drops a still-CONFIRMED override along with the cancelled series.
-func newEventOccChecker(evt event.Event) occChecker {
-	if cancelledRecurringMaster(evt.RecurrenceRule, evt.Status) {
-		return occChecker{}
-	}
-	rs, _ := newEventRRuleSet(evt, false)
-	return occChecker{rs: rs, anchor: evt.StartTime}
-}
-
-// singleEventInstance returns evt as a lone occurrence when it is non-recurring
-// or carries an unparseable RRULE, if its start falls within [from, to).
-func singleEventInstance(evt event.Event, from, to time.Time) []ExpandedEvent {
-	if evt.StartTime.Before(from) || !evt.StartTime.Before(to) {
-		return nil
-	}
-	return []ExpandedEvent{{Event: evt, InstanceTime: evt.StartTime}}
-}
-
-// ExpandEvent generates all occurrences of an event within a date range
-// Returns instances even for non-recurring events (single instance)
-func ExpandEvent(evt event.Event, from, to time.Time) []ExpandedEvent {
-	if cancelledRecurringMaster(evt.RecurrenceRule, evt.Status) {
-		return nil
-	}
-	rs, ok := newEventRRuleSet(evt, true)
-	if !ok {
-		return singleEventInstance(evt, from, to)
-	}
-
-	var instances []ExpandedEvent
-	for _, occ := range rs.between(from, to) {
-		_, isRDate := rs.rdateSet[rdateKey(occ)]
-		instances = append(instances, ExpandedEvent{
-			Event:        evt,
-			InstanceTime: occ.UTC(),
-			IsOverride:   isRDate,
-		})
-	}
-	return instances
-}
-
-// ExpandOption configures ListExpandedEvents behaviour.
-type ExpandOption func(*expandOptions)
-
-type expandOptions struct {
-	skipCategories bool
-}
-
-// SkipCategories omits the batch category load. Use this when the caller
-// does not need Event.Categories (for example alarm checks).
-func SkipCategories() ExpandOption {
-	return func(o *expandOptions) { o.skipCategories = true }
-}
-
-// ListExpandedEvents returns events with their instances in a date range.
-// It uses filtered queries instead of a load of the entire table.
-func (s *Service) ListExpandedEvents(ctx context.Context, from, to time.Time, opts ...ExpandOption) ([]ExpandedEvent, error) {
-	var o expandOptions
-	for _, fn := range opts {
-		fn(&o)
-	}
-	// Non-recurring events in date range.
-	rangeRows, err := s.q.ListEventsByDateRange(ctx, storage.ListEventsByDateRangeParams{
-		StartTime: to.UTC().Format(time.RFC3339),   // start_time < to
-		EndTime:   from.UTC().Format(time.RFC3339), // end_time > from
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// All recurring master events (need full set for expansion).
-	recurRows, err := s.q.ListRecurringEvents(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var results []ExpandedEvent
-
-	for _, row := range rangeRows {
-		// Skip recurring masters (RRULE or RDATE-only) — handled via recurRows below.
-		if row.RecurrenceID == "" && (row.RecurrenceRule != nil || isRDateOnlyMaster(row.Rdates)) {
-			continue
-		}
-		// Skip overrides; they're merged during recurring expansion below.
-		if row.RecurrenceID != "" {
-			continue
-		}
-		evt := event.FromStorage(row)
-		results = append(results, ExpandedEvent{
-			Event:        evt,
-			InstanceTime: evt.StartTime,
-		})
-	}
-
-	// Merge recurring masters with their overrides through the shared engine.
-	// expandedEventKind keeps ExpandedEvent as both the master Model and the
-	// emitted output, so each occurrence's InstanceTime survives (unlike
-	// expandRecurringRows, which folds the occurrence time into Event.Start/End).
-	// The engine batch-fetches overrides and propagates any fetch error, so a
-	// failed lookup never renders a master as if it had none.
-	recurResults, err := expandRecurringRowsBy(ctx, expandedEventKind(s), recurRows, from, to)
-	if err != nil {
-		return nil, err
-	}
-	results = append(results, recurResults...)
-
-	if len(results) > 0 {
-		ids := make([]int64, len(results))
-		for i := range results {
-			ids[i] = results[i].ID
-		}
-
-		if !o.skipCategories {
-			cats, err := s.q.ListCategoriesByEventIDs(ctx, ids)
-			if err == nil {
-				catMap := make(map[int64][]string, len(results))
-				for _, c := range cats {
-					catMap[c.EventID] = append(catMap[c.EventID], c.Category)
-				}
-				for i := range results {
-					if c, ok := catMap[results[i].ID]; ok {
-						results[i].Categories = timeutil.JoinCategoryList(c)
-					}
-				}
-			}
-		}
-
-		atts, err := s.q.ListAttendeesByEventIDs(ctx, ids)
-		if err == nil {
-			attMap := make(map[int64][]model.Attendee, len(results))
-			for _, a := range atts {
-				attMap[a.EventID] = append(attMap[a.EventID], attendeeFromStorage(a))
-			}
-			for i := range results {
-				if a, ok := attMap[results[i].ID]; ok {
-					results[i].Attendees = a
-				}
-			}
-		}
-	}
-
-	sortExpandedEvents(results)
-
-	return results, nil
-}
-
-// sortExpandedEvents orders instances by their local calendar day. It places
-// all-day events before timed events on the same day. Timed events then sort
-// by start time. SQL-level order is not sufficient. Recurring instances
-// are generated in Go.
-func sortExpandedEvents(results []ExpandedEvent) {
-	sort.SliceStable(results, func(i, j int) bool {
-		a, b := results[i], results[j]
-		ad := timeutil.LocalDay(a.InstanceTime)
-		bd := timeutil.LocalDay(b.InstanceTime)
-		if !ad.Equal(bd) {
-			return ad.Before(bd)
-		}
-		if a.AllDay != b.AllDay {
-			return a.AllDay
-		}
-		return a.InstanceTime.Before(b.InstanceTime)
-	})
-}
-
-func attendeeFromStorage(r storage.EventAttendee) model.Attendee {
-	return model.Attendee{
-		ID:            r.ID,
-		EventID:       r.EventID,
-		Email:         r.Email,
-		Name:          storage.NullableToString(r.Name),
-		RSVPStatus:    r.RsvpStatus,
-		Role:          r.Role,
-		Organizer:     r.Organizer == 1,
-		CUType:        storage.NullableToString(r.Cutype),
-		RSVPRequested: strings.EqualFold(storage.NullableToString(r.Rsvp), "TRUE"),
-		SentBy:        storage.NullableToString(r.SentBy),
-		DelegatedTo:   storage.NullableToString(r.DelegatedTo),
-		DelegatedFrom: storage.NullableToString(r.DelegatedFrom),
-		Member:        storage.NullableToString(r.Member),
-		Dir:           storage.NullableToString(r.Dir),
-		Language:      storage.NullableToString(r.Language),
-	}
 }
 
 // recurringKind bundles the entity-specific operations expandRecurringRowsBy
@@ -649,207 +386,6 @@ func expandRecurringRowsBy[Row any, Model any, Inst any](ctx context.Context, k 
 	return result, nil
 }
 
-// expandedEventKind drives expandRecurringRowsBy for ListExpandedEvents, which
-// needs each occurrence's InstanceTime kept. ExpandedEvent serves as both
-// the master Model and the emitted output, so applyInstance is identity. The
-// per-occurrence InstanceTime that expandRecurringRows discards (it collapses
-// the occurrence into Event.Start/End) is kept here.
-func expandedEventKind(s *Service) recurringKind[storage.Event, ExpandedEvent, ExpandedEvent] {
-	return recurringKind[storage.Event, ExpandedEvent, ExpandedEvent]{
-		fromRow: func(r storage.Event) ExpandedEvent {
-			return ExpandedEvent{Event: event.FromStorage(r)}
-		},
-		expand: func(m ExpandedEvent, from, to time.Time) []ExpandedEvent {
-			return ExpandEvent(m.Event, from, to)
-		},
-		instTime:       func(i ExpandedEvent) time.Time { return i.InstanceTime },
-		applyInstance:  func(i ExpandedEvent) ExpandedEvent { return i },
-		uid:            func(r storage.Event) string { return r.Uid },
-		status:         func(r storage.Event) string { return r.Status },
-		recurrenceID:   func(r storage.Event) string { return r.RecurrenceID },
-		overridesByUID: s.eventOverridesByUID,
-		newOccChecker:  func(m ExpandedEvent) occChecker { return newEventOccChecker(m.Event) },
-		emitOverride: func(o storage.Event, from, to time.Time) (ExpandedEvent, bool) {
-			oe := event.FromStorage(o)
-			return ExpandedEvent{Event: oe, InstanceTime: oe.StartTime}, overlapsWindow(oe.StartTime, oe.EndTime, from, to)
-		},
-	}
-}
-
-// expandRecurringRows expands recurring event rows into Event instances with
-// StartTime/EndTime adjusted to each occurrence. For each master, overrides
-// (rows whose RECURRENCE-ID matches) replace the original RRULE instance.
-func (s *Service) expandRecurringRows(ctx context.Context, rows []storage.Event, from, to time.Time) ([]event.Event, error) {
-	k := recurringKind[storage.Event, event.Event, ExpandedEvent]{
-		fromRow:  event.FromStorage,
-		expand:   ExpandEvent,
-		instTime: func(i ExpandedEvent) time.Time { return i.InstanceTime },
-		applyInstance: func(i ExpandedEvent) event.Event {
-			e := i.Event
-			dur := e.EndTime.Sub(e.StartTime)
-			e.StartTime = i.InstanceTime
-			e.EndTime = i.InstanceTime.Add(dur)
-			return e
-		},
-		uid:            func(r storage.Event) string { return r.Uid },
-		status:         func(r storage.Event) string { return r.Status },
-		recurrenceID:   func(r storage.Event) string { return r.RecurrenceID },
-		overridesByUID: s.eventOverridesByUID,
-		newOccChecker:  newEventOccChecker,
-		emitOverride: func(o storage.Event, from, to time.Time) (event.Event, bool) {
-			oe := event.FromStorage(o)
-			return oe, overlapsWindow(oe.StartTime, oe.EndTime, from, to)
-		},
-	}
-	return expandRecurringRowsBy(ctx, k, rows, from, to)
-}
-
-// ListExpandedByDateRange returns non-recurring events in [from,to) merged
-// with expanded instances of recurring event masters. The returned events have
-// StartTime/EndTime adjusted to the instance time and are sorted by StartTime.
-func (s *Service) ListExpandedByDateRange(ctx context.Context, from, to time.Time) ([]event.Event, error) {
-	rangeRows, err := s.q.ListEventsByDateRange(ctx, storage.ListEventsByDateRangeParams{
-		StartTime: to.UTC().Format(time.RFC3339),   // start_time < to
-		EndTime:   from.UTC().Format(time.RFC3339), // end_time > from
-	})
-	if err != nil {
-		return nil, err
-	}
-	// Keep only non-recurring, non-RDATE-only masters from the date-range results.
-	// RDATE-only masters (RecurrenceRule IS NULL but Rdates non-empty) must follow
-	// the recurring-expansion path below, not be emitted as singletons here.
-	var result []event.Event
-	for _, row := range rangeRows {
-		if row.RecurrenceRule == nil && row.RecurrenceID == "" && !isRDateOnlyMaster(row.Rdates) {
-			result = append(result, event.FromStorage(row))
-		}
-	}
-
-	recurringRows, err := s.q.ListRecurringEvents(ctx)
-	if err != nil {
-		return nil, err
-	}
-	expanded, err := s.expandRecurringRows(ctx, recurringRows, from, to)
-	if err != nil {
-		return nil, err
-	}
-	result = append(result, expanded...)
-
-	s.populateEventCategories(ctx, result)
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].StartTime.Before(result[j].StartTime)
-	})
-	return result, nil
-}
-
-// ExportFilterParams holds filters for ICS export of recurring masters.
-type ExportFilterParams struct {
-	CalendarID int64
-	Category   string
-	Status     string
-	From       time.Time
-	To         time.Time
-}
-
-// rangeBoundUTC formats a date-range query bound as an RFC3339 string in UTC,
-// or "" for the zero time. Normalize to UTC because these bounds are compared
-// lexically against the UTC-stored start/end strings (issue #305). A non-UTC
-// offset left in the formatted string breaks that comparison near window edges.
-func rangeBoundUTC(t time.Time) string {
-	if t.IsZero() {
-		return ""
-	}
-	return t.UTC().Format(time.RFC3339)
-}
-
-// ExportExpandedByDateRange returns recurring event masters (not expanded
-// instances) that have at least one occurrence in [from,to). It merges them
-// with non-recurring events whose start_time is in range. This is for ICS
-// export. Emit the master VEVENT with RRULE, not individual instances. All
-// filters (calendar, category, status) are applied at the SQL level.
-func (s *Service) ExportExpandedByDateRange(ctx context.Context, p ExportFilterParams) ([]event.Event, error) {
-	fromStr := rangeBoundUTC(p.From)
-	toStr := rangeBoundUTC(p.To)
-
-	rangeRows, err := s.q.ListEventsForExport(ctx, storage.EventFilterParams{
-		CalendarID:   p.CalendarID,
-		FromTime:     fromStr,
-		ToTime:       toStr,
-		Category:     p.Category,
-		FilterStatus: p.Status,
-	})
-	if err != nil {
-		return nil, err
-	}
-	var result []event.Event
-	seen := make(map[int64]bool)
-	for _, row := range rangeRows {
-		if row.RecurrenceRule == nil {
-			result = append(result, event.FromStorage(row))
-			seen[row.ID] = true
-		}
-	}
-
-	recurringRows, err := s.q.ListRecurringEventsFiltered(ctx, storage.EventFilterParams{
-		CalendarID:   p.CalendarID,
-		FilterStatus: p.Status,
-		Category:     p.Category,
-	})
-	if err != nil {
-		return nil, err
-	}
-	for _, row := range recurringRows {
-		if seen[row.ID] {
-			continue
-		}
-		evt := event.FromStorage(row)
-		// Export must emit a cancelled master (STATUS:CANCELLED is how a
-		// downstream client is told to drop the series), so the in-range probe
-		// ignores the cancelled-expansion guard — unlike display, which hides it.
-		probe := evt
-		probe.Status = ""
-		if len(ExpandEvent(probe, p.From, p.To)) > 0 {
-			result = append(result, evt)
-		}
-	}
-
-	s.populateEventCategories(ctx, result)
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].StartTime.Before(result[j].StartTime)
-	})
-	return result, nil
-}
-
-func todoFromRow(row storage.Todo) todo.Todo {
-	return todo.Todo{
-		ID:              row.ID,
-		UID:             row.Uid,
-		CalendarID:      row.CalendarID,
-		Summary:         row.Summary,
-		Description:     storage.NullableToString(row.Description),
-		Location:        storage.NullableToString(row.Location),
-		DueDate:         storage.NullableToString(row.DueDate),
-		StartDate:       storage.NullableToString(row.StartDate),
-		Duration:        storage.NullableToString(row.Duration),
-		CompletedAt:     storage.NullableToString(row.CompletedAt),
-		PercentComplete: row.PercentComplete,
-		Status:          row.Status,
-		Priority:        row.Priority,
-		Class:           row.Class,
-		URL:             storage.NullableToString(row.Url),
-		RecurrenceRule:  storage.NullableToString(row.RecurrenceRule),
-		Timezone:        storage.NullableToString(row.Timezone),
-		Sequence:        row.Sequence,
-		ExDates:         storage.NullableToString(row.Exdates),
-		RDates:          storage.NullableToString(row.Rdates),
-		RecurrenceID:    row.RecurrenceID,
-		Geo:             storage.NullableToString(row.Geo),
-		DtStamp:         storage.NullableToString(row.Dtstamp),
-		CreatedAt:       timeutil.ParseDateTime(row.CreatedAt),
-		UpdatedAt:       timeutil.ParseDateTime(row.UpdatedAt),
-	}
-}
-
 // populateCategories batch-loads categories for items and assigns the joined
 // category string to each via setCats. idOf yields an item's primary key.
 // fetch loads the join rows for a set of ids. rowCat splits a join row into
@@ -886,177 +422,6 @@ func populateCategories[T any, R any](
 	}
 }
 
-// populateEventAttendees attaches guests in one batched query. event list JSON
-// is the Chroncal Bar agenda payload; without this it omits invitations that
-// event get already hydrates. Generated occurrences keep the master ID, so they
-// receive the master's attendee rows. A load error leaves Attendees unset,
-// matching populateEventCategories.
-func (s *Service) populateEventAttendees(ctx context.Context, events []event.Event) {
-	if len(events) == 0 {
-		return
-	}
-	ids := make([]int64, len(events))
-	for i := range events {
-		ids[i] = events[i].ID
-	}
-	atts, err := s.q.ListAttendeesByEventIDs(ctx, ids)
-	if err != nil {
-		return
-	}
-	attMap := make(map[int64][]model.Attendee, len(events))
-	for _, a := range atts {
-		attMap[a.EventID] = append(attMap[a.EventID], attendeeFromStorage(a))
-	}
-	for i := range events {
-		if a, ok := attMap[events[i].ID]; ok {
-			events[i].Attendees = a
-		}
-	}
-}
-
-func (s *Service) populateEventCategories(ctx context.Context, events []event.Event) {
-	populateCategories(ctx, events,
-		func(e event.Event) int64 { return e.ID },
-		s.q.ListCategoriesByEventIDs,
-		func(r storage.EventCategory) (int64, string) { return r.EventID, r.Category },
-		func(e *event.Event, joined string) { e.Categories = joined },
-	)
-}
-
-func (s *Service) populateTodoCategories(ctx context.Context, todos []todo.Todo) {
-	populateCategories(ctx, todos,
-		func(t todo.Todo) int64 { return t.ID },
-		s.q.ListCategoriesByTodoIDs,
-		func(r storage.TodoCategory) (int64, string) { return r.TodoID, r.Category },
-		func(t *todo.Todo, joined string) { t.Categories = joined },
-	)
-}
-
-// todoAnchor returns a todo's recurrence anchor: DTSTART if present, else DUE.
-func todoAnchor(td todo.Todo) time.Time {
-	anchor := td.ParseStartDate()
-	if anchor.IsZero() {
-		anchor = td.ParseDueDate()
-	}
-	return anchor
-}
-
-// todoDuration is the START->DUE span used to keep an occurrence that spans the window. A
-// due-only (point) todo has none.
-func todoDuration(td todo.Todo) time.Duration {
-	if start := td.ParseStartDate(); !start.IsZero() {
-		if due := td.ParseDueDate(); !due.IsZero() && due.After(start) {
-			return due.Sub(start)
-		}
-	}
-	return 0
-}
-
-// newTodoRRuleSet parses td's RRULE into a reusable rruleSet. ok is false when
-// the todo has no anchor (so it cannot be expanded).
-func newTodoRRuleSet(td todo.Todo, includeExDates bool) (rruleSet, bool) {
-	anchor := todoAnchor(td)
-	if anchor.IsZero() {
-		return rruleSet{}, false
-	}
-	return newRRuleSet(td.RecurrenceRule, td.Timezone, anchor,
-		todoDuration(td), td.ParseExDates(), td.ParseRDates(), includeExDates)
-}
-
-// newTodoOccChecker builds a reusable orphan-detection checker for a recurring
-// todo master. A cancelled series matches nothing (see newEventOccChecker).
-func newTodoOccChecker(td todo.Todo) occChecker {
-	if cancelledRecurringMaster(td.RecurrenceRule, td.Status) {
-		return occChecker{}
-	}
-	rs, _ := newTodoRRuleSet(td, false)
-	return occChecker{rs: rs, anchor: todoAnchor(td)}
-}
-
-// singleTodoInstance returns td as a lone occurrence (non-recurring or
-// unparseable RRULE) if its anchor falls within [from, to).
-func singleTodoInstance(td todo.Todo, from, to time.Time) []ExpandedTodo {
-	anchor := todoAnchor(td)
-	if anchor.IsZero() || anchor.Before(from) || !anchor.Before(to) {
-		return nil
-	}
-	return []ExpandedTodo{{Todo: td, InstanceTime: anchor}}
-}
-
-// ExpandTodo generates all occurrences of a todo within a date range.
-// The anchor date is DTSTART if present, else DUE. For non-recurring todos
-// a single instance is returned if the anchor falls in range.
-func ExpandTodo(td todo.Todo, from, to time.Time) []ExpandedTodo {
-	// A cancelled recurring master has no occurrences (see cancelledRecurringMaster).
-	if cancelledRecurringMaster(td.RecurrenceRule, td.Status) {
-		return nil
-	}
-	rs, ok := newTodoRRuleSet(td, true)
-	if !ok {
-		return singleTodoInstance(td, from, to)
-	}
-
-	var instances []ExpandedTodo
-	for _, occ := range rs.between(from, to) {
-		_, isRDate := rs.rdateSet[rdateKey(occ)]
-		instances = append(instances, ExpandedTodo{
-			Todo:         td,
-			InstanceTime: occ.UTC(),
-			IsOverride:   isRDate,
-		})
-	}
-	return instances
-}
-
-// expandRecurringTodoRows expands recurring todo rows into Todo instances with
-// DueDate/StartDate adjusted to each occurrence. For each master, overrides
-// (rows with a match of RECURRENCE-ID) replace the original RRULE instance.
-func (s *Service) expandRecurringTodoRows(ctx context.Context, rows []storage.Todo, from, to time.Time) ([]todo.Todo, error) {
-	k := recurringKind[storage.Todo, todo.Todo, ExpandedTodo]{
-		fromRow:  todoFromRow,
-		expand:   ExpandTodo,
-		instTime: func(i ExpandedTodo) time.Time { return i.InstanceTime },
-		applyInstance: func(i ExpandedTodo) todo.Todo {
-			t := i.Todo
-			anchor := t.ParseStartDate()
-			if anchor.IsZero() {
-				anchor = t.ParseDueDate()
-			}
-			if !anchor.IsZero() {
-				offset := i.InstanceTime.Sub(anchor)
-				t.DueDate = shiftDateString(t.DueDate, t.ParseDueDate(), offset)
-				t.StartDate = shiftDateString(t.StartDate, t.ParseStartDate(), offset)
-			}
-			return t
-		},
-		uid:            func(r storage.Todo) string { return r.Uid },
-		status:         func(r storage.Todo) string { return r.Status },
-		recurrenceID:   func(r storage.Todo) string { return r.RecurrenceID },
-		overridesByUID: s.todoOverridesByUID,
-		newOccChecker:  newTodoOccChecker,
-		emitOverride: func(o storage.Todo, from, to time.Time) (todo.Todo, bool) {
-			ot := todoFromRow(o)
-			anchor := ot.ParseStartDate()
-			if anchor.IsZero() {
-				anchor = ot.ParseDueDate()
-			}
-			if anchor.IsZero() {
-				// No datable anchor: fall back to the replaced slot for the
-				// window check. An unparseable recurrence_id leaves anchor zero,
-				// which fails keepOccurrence and is dropped (the orphan probe
-				// that follows would drop it too).
-				anchor, _ = timeutil.ParseRecurrenceID(o.RecurrenceID)
-			}
-			// Keep on [START, DUE) interval overlap (honoring todoDuration), so a
-			// multi-day override whose START precedes the window but whose DUE
-			// falls inside it is not dropped -- matching the master-occurrence
-			// path (keepOccurrence/between) and the event override path.
-			return ot, keepOccurrence(anchor, todoDuration(ot), from, to)
-		},
-	}
-	return expandRecurringRowsBy(ctx, k, rows, from, to)
-}
-
 // shiftDateString returns value advanced by offset. It keeps its date-only or
 // RFC 3339 representation. It returns value unchanged when it is empty or its
 // parsed form (parsed) is zero. A blank or unparseable field is then left intact.
@@ -1069,413 +434,4 @@ func shiftDateString(value string, parsed time.Time, offset time.Duration) strin
 		return shifted.Format("2006-01-02")
 	}
 	return shifted.Format(time.RFC3339)
-}
-
-// ListExpandedTodosByDueDateRange returns non-recurring todos in [from,to)
-// merged with expanded instances of recurring todo masters.
-func (s *Service) ListExpandedTodosByDueDateRange(ctx context.Context, from, to time.Time) ([]todo.Todo, error) {
-	fromStr := from.Format("2006-01-02")
-	toStr := to.Format("2006-01-02")
-	rangeRows, err := s.q.ListTodosByDueDateRange(ctx, storage.ListTodosByDueDateRangeParams{
-		DueDate:   &fromStr,
-		DueDate_2: &toStr,
-	})
-	if err != nil {
-		return nil, err
-	}
-	// Keep only non-recurring, non-RDATE-only masters from the due-date results.
-	var result []todo.Todo
-	for _, row := range rangeRows {
-		if row.RecurrenceRule == nil && row.RecurrenceID == "" && !isRDateOnlyMaster(row.Rdates) {
-			result = append(result, todoFromRow(row))
-		}
-	}
-
-	recurringRows, err := s.q.ListRecurringTodos(ctx)
-	if err != nil {
-		return nil, err
-	}
-	expanded, err := s.expandRecurringTodoRows(ctx, recurringRows, from, to)
-	if err != nil {
-		return nil, err
-	}
-	result = append(result, expanded...)
-
-	s.populateTodoCategories(ctx, result)
-	sort.Slice(result, func(i, j int) bool {
-		di := result[i].ParseDueDate()
-		dj := result[j].ParseDueDate()
-		if di.IsZero() {
-			return false
-		}
-		if dj.IsZero() {
-			return true
-		}
-		return di.Before(dj)
-	})
-	return result, nil
-}
-
-// EventListParams holds composable filters for event lists.
-type EventListParams struct {
-	CalendarID     int64
-	Status         string
-	Category       string
-	From           time.Time
-	To             time.Time
-	IncludeDeleted bool
-}
-
-// ListFilteredEvents returns events that match all supplied filters. Calendar,
-// status, and date-range filters compose freely. When a date range is
-// provided, recurring events are expanded within it, with overrides applied.
-// Otherwise recurring masters are returned as-is. This matches the
-// todo/journal contract.
-func (s *Service) ListFilteredEvents(ctx context.Context, p EventListParams) ([]event.Event, error) {
-	fromStr := rangeBoundUTC(p.From)
-	toStr := rangeBoundUTC(p.To)
-	hasRange := fromStr != "" || toStr != ""
-
-	rangeRows, err := s.q.ListEventsFiltered(ctx, storage.EventFilterParams{
-		CalendarID:     p.CalendarID,
-		FilterStatus:   p.Status,
-		Category:       p.Category,
-		FromTime:       fromStr,
-		ToTime:         toStr,
-		IncludeDeleted: p.IncludeDeleted,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	var result []event.Event
-	for _, row := range rangeRows {
-		result = append(result, event.FromStorage(row))
-	}
-
-	recurringRows, err := s.q.ListRecurringEventsFiltered(ctx, storage.EventFilterParams{
-		CalendarID:     p.CalendarID,
-		FilterStatus:   p.Status,
-		Category:       p.Category,
-		IncludeDeleted: p.IncludeDeleted,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if hasRange {
-		expanded, err := s.expandRecurringRows(ctx, recurringRows, p.From, p.To)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, expanded...)
-	} else {
-		for _, row := range recurringRows {
-			result = append(result, event.FromStorage(row))
-		}
-	}
-
-	s.populateEventCategories(ctx, result)
-	s.populateEventAttendees(ctx, result)
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].StartTime.Before(result[j].StartTime)
-	})
-	return result, nil
-}
-
-// TodoListParams holds composable filters for todo lists.
-type TodoListParams struct {
-	CalendarID    int64
-	Status        string
-	HideCompleted bool
-	From          time.Time
-	To            time.Time
-	// IncludeDeleted, when true, returns soft-deleted todos alongside live
-	// rows. Default (false) hides them, matching the live-query contract
-	// every other service method honors.
-	IncludeDeleted bool
-}
-
-// ListFilteredTodos returns todos that match all supplied filters. When a date
-// range is provided, recurring todos are expanded. Otherwise master entries
-// are returned as-is.
-func (s *Service) ListFilteredTodos(ctx context.Context, p TodoListParams) ([]todo.Todo, error) {
-	hideCompleted := int64(0)
-	if p.HideCompleted {
-		hideCompleted = 1
-	}
-
-	fromStr := ""
-	toStr := ""
-	hasRange := !p.From.IsZero() || !p.To.IsZero()
-	if !p.From.IsZero() {
-		fromStr = p.From.Format("2006-01-02")
-	}
-	if !p.To.IsZero() {
-		toStr = p.To.Format("2006-01-02")
-	}
-
-	rows, err := s.q.ListTodosFiltered(ctx, storage.ListTodosFilteredParams{
-		CalendarID:     p.CalendarID,
-		FilterStatus:   p.Status,
-		HideCompleted:  hideCompleted,
-		FromDate:       fromStr,
-		ToDate:         toStr,
-		IncludeDeleted: p.IncludeDeleted,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	var result []todo.Todo
-	for _, row := range rows {
-		result = append(result, todoFromRow(row))
-	}
-
-	recurringRows, err := s.q.ListRecurringTodosFiltered(ctx, storage.ListRecurringTodosFilteredParams{
-		CalendarID:     p.CalendarID,
-		FilterStatus:   p.Status,
-		HideCompleted:  hideCompleted,
-		IncludeDeleted: p.IncludeDeleted,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if hasRange {
-		expanded, err := s.expandRecurringTodoRows(ctx, recurringRows, p.From, p.To)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, expanded...)
-	} else {
-		for _, row := range recurringRows {
-			result = append(result, todoFromRow(row))
-		}
-	}
-
-	s.populateTodoCategories(ctx, result)
-	sort.Slice(result, func(i, j int) bool {
-		di := result[i].ParseDueDate()
-		dj := result[j].ParseDueDate()
-		if di.IsZero() {
-			return false
-		}
-		if dj.IsZero() {
-			return true
-		}
-		return di.Before(dj)
-	})
-	return result, nil
-}
-
-// ── Journal recurrence ────────────────────────────────────────────────
-
-// ExpandedJournal represents a single occurrence of a (possibly recurring) journal.
-type ExpandedJournal struct {
-	journal.Journal
-	InstanceTime time.Time
-	IsOverride   bool
-}
-
-// JournalListParams holds composable filters for journal lists.
-type JournalListParams struct {
-	CalendarID int64
-	Status     string
-	// HideCancelled, when true, omits CANCELLED journals. Default (false)
-	// returns every status, matching the iCal model where a cancelled
-	// journal is still a real row the caller may want to see.
-	HideCancelled bool
-	From          time.Time
-	To            time.Time
-	// IncludeDeleted, when true, returns soft-deleted journals alongside
-	// live rows. Default (false) hides them.
-	IncludeDeleted bool
-}
-
-func journalFromRow(row storage.Journal) journal.Journal {
-	return journal.Journal{
-		ID:             row.ID,
-		UID:            row.Uid,
-		CalendarID:     row.CalendarID,
-		Summary:        row.Summary,
-		Description:    storage.NullableToString(row.Description),
-		StartDate:      storage.NullableToString(row.StartDate),
-		Status:         row.Status,
-		Class:          row.Class,
-		URL:            storage.NullableToString(row.Url),
-		RecurrenceRule: storage.NullableToString(row.RecurrenceRule),
-		Timezone:       storage.NullableToString(row.Timezone),
-		Sequence:       row.Sequence,
-		ExDates:        storage.NullableToString(row.Exdates),
-		RDates:         storage.NullableToString(row.Rdates),
-		RecurrenceID:   row.RecurrenceID,
-		DtStamp:        storage.NullableToString(row.Dtstamp),
-		CreatedAt:      timeutil.ParseDateTime(row.CreatedAt),
-		UpdatedAt:      timeutil.ParseDateTime(row.UpdatedAt),
-	}
-}
-
-func (s *Service) populateJournalCategories(ctx context.Context, journals []journal.Journal) {
-	populateCategories(ctx, journals,
-		func(j journal.Journal) int64 { return j.ID },
-		s.q.ListCategoriesByJournalIDs,
-		func(r storage.JournalCategory) (int64, string) { return r.JournalID, r.Category },
-		func(j *journal.Journal, joined string) { j.Categories = joined },
-	)
-}
-
-// newJournalRRuleSet parses j's RRULE into a reusable rruleSet. ok is false when
-// the journal has no start date (so it cannot be expanded). Journals are points
-// in time, so the set carries no duration.
-func newJournalRRuleSet(j journal.Journal, includeExDates bool) (rruleSet, bool) {
-	anchor := j.ParseStartDate()
-	if anchor.IsZero() {
-		return rruleSet{}, false
-	}
-	return newRRuleSet(j.RecurrenceRule, j.Timezone, anchor, 0,
-		j.ParseExDates(), j.ParseRDates(), includeExDates)
-}
-
-// newJournalOccChecker builds a reusable orphan-detection checker for a
-// recurring journal master. A cancelled series matches nothing (see
-// newEventOccChecker).
-func newJournalOccChecker(j journal.Journal) occChecker {
-	if cancelledRecurringMaster(j.RecurrenceRule, j.Status) {
-		return occChecker{}
-	}
-	rs, _ := newJournalRRuleSet(j, false)
-	return occChecker{rs: rs, anchor: j.ParseStartDate()}
-}
-
-// singleJournalInstance returns j as a lone occurrence (non-recurring or
-// unparseable RRULE) if its start date falls within [from, to).
-func singleJournalInstance(j journal.Journal, from, to time.Time) []ExpandedJournal {
-	anchor := j.ParseStartDate()
-	if anchor.IsZero() || anchor.Before(from) || !anchor.Before(to) {
-		return nil
-	}
-	return []ExpandedJournal{{Journal: j, InstanceTime: anchor}}
-}
-
-// ExpandJournal generates all occurrences of a journal within a date range.
-func ExpandJournal(j journal.Journal, from, to time.Time) []ExpandedJournal {
-	// A cancelled recurring master has no occurrences (see cancelledRecurringMaster).
-	if cancelledRecurringMaster(j.RecurrenceRule, j.Status) {
-		return nil
-	}
-	rs, ok := newJournalRRuleSet(j, true)
-	if !ok {
-		return singleJournalInstance(j, from, to)
-	}
-
-	var instances []ExpandedJournal
-	for _, occ := range rs.between(from, to) {
-		_, isRDate := rs.rdateSet[rdateKey(occ)]
-		instances = append(instances, ExpandedJournal{Journal: j, InstanceTime: occ.UTC(), IsOverride: isRDate})
-	}
-	return instances
-}
-
-func (s *Service) expandRecurringJournalRows(ctx context.Context, rows []storage.Journal, from, to time.Time) ([]journal.Journal, error) {
-	k := recurringKind[storage.Journal, journal.Journal, ExpandedJournal]{
-		fromRow:  journalFromRow,
-		expand:   ExpandJournal,
-		instTime: func(i ExpandedJournal) time.Time { return i.InstanceTime },
-		applyInstance: func(i ExpandedJournal) journal.Journal {
-			jj := i.Journal
-			if anchor := jj.ParseStartDate(); !anchor.IsZero() {
-				jj.StartDate = shiftDateString(jj.StartDate, anchor, i.InstanceTime.Sub(anchor))
-			}
-			return jj
-		},
-		uid:            func(r storage.Journal) string { return r.Uid },
-		status:         func(r storage.Journal) string { return r.Status },
-		recurrenceID:   func(r storage.Journal) string { return r.RecurrenceID },
-		overridesByUID: s.journalOverridesByUID,
-		newOccChecker:  newJournalOccChecker,
-		emitOverride: func(o storage.Journal, from, to time.Time) (journal.Journal, bool) {
-			oj := journalFromRow(o)
-			anchor := oj.ParseStartDate()
-			if anchor.IsZero() {
-				// No datable anchor: fall back to the replaced slot for the
-				// window check. An unparseable recurrence_id leaves anchor zero,
-				// which fails inWindow and is dropped (the orphan probe that
-				// follows would drop it too).
-				anchor, _ = timeutil.ParseRecurrenceID(o.RecurrenceID)
-			}
-			return oj, inWindow(anchor, from, to)
-		},
-	}
-	return expandRecurringRowsBy(ctx, k, rows, from, to)
-}
-
-// ListFilteredJournals returns journals that match all supplied filters. When a
-// date range is provided, recurring journals are expanded. Otherwise master
-// entries are returned as-is.
-func (s *Service) ListFilteredJournals(ctx context.Context, p JournalListParams) ([]journal.Journal, error) {
-	hideCancelled := int64(0)
-	if p.HideCancelled {
-		hideCancelled = 1
-	}
-
-	fromStr := ""
-	toStr := ""
-	hasRange := !p.From.IsZero() || !p.To.IsZero()
-	if !p.From.IsZero() {
-		fromStr = p.From.Format("2006-01-02")
-	}
-	if !p.To.IsZero() {
-		toStr = p.To.Format("2006-01-02")
-	}
-
-	rows, err := s.q.ListJournalsFiltered(ctx, storage.ListJournalsFilteredParams{
-		CalendarID:     p.CalendarID,
-		FilterStatus:   p.Status,
-		HideCancelled:  hideCancelled,
-		FromDate:       fromStr,
-		ToDate:         toStr,
-		IncludeDeleted: p.IncludeDeleted,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	var result []journal.Journal
-	for _, row := range rows {
-		result = append(result, journalFromRow(row))
-	}
-
-	recurringRows, err := s.q.ListRecurringJournalsFiltered(ctx, storage.ListRecurringJournalsFilteredParams{
-		CalendarID:     p.CalendarID,
-		FilterStatus:   p.Status,
-		HideCancelled:  hideCancelled,
-		IncludeDeleted: p.IncludeDeleted,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if hasRange {
-		expanded, err := s.expandRecurringJournalRows(ctx, recurringRows, p.From, p.To)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, expanded...)
-	} else {
-		for _, row := range recurringRows {
-			result = append(result, journalFromRow(row))
-		}
-	}
-
-	s.populateJournalCategories(ctx, result)
-	sort.Slice(result, func(i, j int) bool {
-		di := result[i].ParseStartDate()
-		dj := result[j].ParseStartDate()
-		if di.IsZero() {
-			return false
-		}
-		if dj.IsZero() {
-			return true
-		}
-		return di.Before(dj)
-	})
-	return result, nil
 }
